server/pyproject.toml
[tool.poetry]
name = "battlefield"
version = "0.1.0"
description = "Testing Quart and Websockets"
authors = ["CristhianMotoche <cristhian.motoche@gmail.com>"]
license = "WTFPL"

[tool.poetry.dependencies]
python = "^3.7"
Quart = "^0.9.1"
tortoise-orm = "^0.13.10"
aiosqlite = "^0.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.1"
pytest-asyncio = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
